+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxAuxData.hs
@@ -34,7 +34,6 @@ module Cardano.Ledger.Alonzo.TxAuxData (
   AlonzoEraTxAuxData (..),
   AlonzoTxAuxDataRaw,
   mkAlonzoTxAuxData,
-  AuxiliaryDataHash (..),
   hashAlonzoTxAuxData,
   validateAlonzoTxAuxData,
   getAlonzoTxAuxDataScripts,
@@ -56,7 +55,6 @@ import Cardano.Ledger.Alonzo.Scripts (
   plutusScriptLanguage,
   validScript,
  )
-import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
 import Cardano.Ledger.BaseTypes (ProtVer)
 import Cardano.Ledger.Binary (
   Annotator (..),
@@ -263,8 +261,6 @@ instance EraTxAuxData AlonzoEra where
         , atadrPlutus = mempty
         }
 
-  hashTxAuxData = hashAlonzoTxAuxData
-
   validateTxAuxData = validateAlonzoTxAuxData
 
 metadataAlonzoTxAuxDataL :: Era era => Lens' (AlonzoTxAuxData era) (Map Word64 Metadatum)
@@ -274,8 +270,9 @@ metadataAlonzoTxAuxDataL =
 hashAlonzoTxAuxData ::
   HashAnnotated x EraIndependentTxAuxData =>
   x ->
-  AuxiliaryDataHash
-hashAlonzoTxAuxData x = AuxiliaryDataHash (hashAnnotated x)
+  TxAuxDataHash
+hashAlonzoTxAuxData x = TxAuxDataHash (hashAnnotated x)
+{-# DEPRECATED hashAlonzoTxAuxData "In favor of `hashTxAuxData`" #-}
 
 validateAlonzoTxAuxData ::
   (AlonzoEraScript era, Script era ~ AlonzoScript era) =>